chore: complete integrity review batch fixes (#205)
- Extract mmmu/jmmmu shared logic to mmmu_utils.py
- Extract column rename utility in make_leaderboard.py
- Move model display names to metadata.py
- Fix task_id/task_dir naming in dashboard.py and make_leaderboard.py
- Move empty prediction check before LLM call
- Add MAX_IMAGES_PER_QUESTION constant
- Add deprecation warning to examples/utils.py

Refs #194
